backend/requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
pyjwt[crypto]==2.10.1
passlib[bcrypt]==1.7.4
python-dotenv==1.1.1
sqlmodel==0.0.24
python-multipart==0.0.20
cryptography==45.0.4
langchain==0.3.26
langchain-openai==0.3.27